requirements.txt
bert4keras
numpy
gensim
tensorflow-probability==0.5.0
ipdb
